loading progress is now per ui element
before, when loading thumbnails, the entire ui was blocked by
a progress message.

now, each directive which loads something will have a separate loading
indicator.

also, fixed the following:
before, when a new directory was chosen and it had no images, the previous
image was still shown.
Now the image is cleared before loading the new directory, and no image
will be shown if the directory contains no images.